Changed name of itoa() function

+++ src/ft_itoa_unsigned_base.c
@@ -6,18 +6,18 @@
 /*   By: fcatina <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/02 23:22:44 by fcatina           #+#    #+#             */
-/*   Updated: 2020/02/02 23:23:57 by fcatina          ###   ########.fr       */
+/*   Updated: 2020/02/22 07:24:14 by slisandr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libftprintf.h"
 
-char	*ft_itoa_unsigned_base(unsigned int value, int base, char c)
+char	*ft_itoa_unsigned_base(unsigned long long value, int base, char c)
 {
-	unsigned int	size;
-	unsigned int	tmp;
-	char			*tab;
-	char			*str;
+	unsigned long long	size;
+	unsigned long long	tmp;
+	char				*tab;
+	char				*str;
 
 	tmp = value;
 	size = 0;
